requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Pillow>6.0<7.0
djangorestframework
djoser